feat: refactor core logic
